flask-honeypot/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
ipinfo==4.4.3
user-agents==2.2.0
gunicorn
pytest